continuum/requirements.txt
Django==4.2.4
psycopg2-binary==2.9.6
gunicorn==21.2.0

authlib==1.0.1
djangorestframework-simplejwt[crypto]==5.2.2
django-cors-headers
djangorestframework
django-filter
requests ~= 2.31
django-redis==5.3.0
django-guardian==2.4.0
sentry-sdk[django]==1.44.1
boto3==1.34.79
django-ses==3.5.2
openai==1.16.2

celery==5.2.7
redis==4.4.0